+++ addOns/fuzz/src/main/java/org/zaproxy/zap/extension/fuzz/messagelocations/MultipleMessageLocationsBreadthFirstReplacer.java
@@ -44,9 +44,6 @@ public class MultipleMessageLocationsBreadthFirstReplacer<T extends Message>
     private boolean initialised;
     private boolean setup;
 
-    private int tailIndex;
-    private MessageLocationReplacementGenerator<?, ?> tail;
-
     private long numberOfReplacements;
 
     @Override
@@ -65,21 +62,24 @@ public class MultipleMessageLocationsBreadthFirstReplacer<T extends Message>
         listCurrentReplacements =
                 new MessageLocationReplacement<?>[messageLocationReplacementGenerator.size()];
 
+        numberOfReplacements = MessageLocationReplacementGenerator.UNKNOWN_NUMBER_OF_REPLACEMENTS;
         replacementGenerators = new ArrayList<>(messageLocationReplacementGenerator.size());
         for (MessageLocationReplacementGenerator<?, ?> mlr : messageLocationReplacementGenerator) {
             if (mlr.hasNext()) {
                 long replacements = mlr.getNumberOfReplacements();
                 if (replacements
                         != MessageLocationReplacementGenerator.UNKNOWN_NUMBER_OF_REPLACEMENTS) {
-                    numberOfReplacements *= replacements;
+                    if (numberOfReplacements
+                            == MessageLocationReplacementGenerator.UNKNOWN_NUMBER_OF_REPLACEMENTS) {
+                        numberOfReplacements = replacements;
+                    } else {
+                        numberOfReplacements = Math.min(numberOfReplacements, replacements);
+                    }
                 }
                 replacementGenerators.add(mlr);
             }
         }
-        numberOfReplacements = 0;
 
-        tailIndex = replacementGenerators.size() - 1;
-        tail = replacementGenerators.get(tailIndex);
         initialised = true;
         setup = true;
     }
@@ -91,36 +91,22 @@ public class MultipleMessageLocationsBreadthFirstReplacer<T extends Message>
 
     @Override
     public boolean hasNext() {
-        for (int i = tailIndex; i >= 0; i--) {
-            if (replacementGenerators.get(i).hasNext()) {
-                return true;
+        for (MessageLocationReplacementGenerator<?, ?> generator : replacementGenerators) {
+            if (!generator.hasNext()) {
+                return false;
             }
         }
-        return false;
+        return !replacementGenerators.isEmpty();
     }
 
     @Override
-    public T next() throws InvalidMessageException {
+    public T next() throws ReplacementException, InvalidMessageException {
         if (setup) {
             setup();
             setup = false;
+        } else {
+            advanceAll();
         }
-
-        if (!tail.hasNext()) {
-            tail.reset();
-
-            for (int i = tailIndex - 1; i >= 0; i--) {
-                if (replacementGenerators.get(i).hasNext()) {
-                    listCurrentReplacements[i] = replacementGenerators.get(i).next();
-                    break;
-                }
-
-                replacementGenerators.get(i).reset();
-                listCurrentReplacements[i] = replacementGenerators.get(i).next();
-            }
-        }
-
-        listCurrentReplacements[tailIndex] = tail.next();
 
         currentReplacements.clear();
         currentReplacements.addAll(Arrays.asList(listCurrentReplacements));
@@ -129,10 +115,12 @@ public class MultipleMessageLocationsBreadthFirstReplacer<T extends Message>
     }
 
     private void setup() {
-        for (int i = 0; i < tailIndex; i++) {
-            if (replacementGenerators.get(i).hasNext()) {
-                listCurrentReplacements[i] = replacementGenerators.get(i).next();
-            }
+        advanceAll();
+    }
+
+    private void advanceAll() {
+        for (int i = 0; i < replacementGenerators.size(); i++) {
+            listCurrentReplacements[i] = replacementGenerators.get(i).next();
         }
     }